Fold abi.rs into capi.rs
